Move some base data structures in audio package from __init__.py to base.py.
